separar las funciones de los heroes
