+++ index.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const heroes = require('./heroesDB.json');
+const actualizarHeroe = require('./actualizarHeroe.js');
+const agregarNuevoHeroe = require('./agregarNuevoHeroe.js');
+const calcularTiempoDeDescanso = require('./calcularTiempoDeDescanso.js');
+const compararEdad = require('./compararEdad.js');
+const compararEditorial = require('./compararEditorial');
+const compararRaza = require('./compararRaza.js');
+const eliminarHeroe = require('./eliminarHeroe.js');
+const encontrarHeroe = require('./encontrarHeroe.js');
+const guardarLista = require('./guardarLista');
+
+console.log('------------------------------');
+console.log(calcularTiempoDeDescanso(4));
